+++ tests/cases/go/cancel-with-replace.ts
@@ -7,13 +7,12 @@ export default {
   msg:
     "pop is cancelled if there is a replace before pending response finishes",
   async: true,
-  fn: function({ history, pathname, resolve }: AsyncTestCaseArgs) {
+  fn: function({ history, resolve }: AsyncTestCaseArgs) {
     function initialRouter(pending) {
       pending.finish();
     }
     let cancelGo;
     const goRouter = ignoreFirstCall(function(pending) {
-      expect(pathname()).toBe("/four");
       cancelGo = pending.cancel;
       // trigger a replace call and don't resolve the go
       history.respondWith(replaceRouter);