+++ packages/in-memory/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+- Remove `history.locations` and `history.index` values.
 - `go` emulates browser behavior by setting `history.index` immediately instead of when the navigation is finished. This means that a subsequent `go` call will be based on the pending `index`, not the index of the current `history.location`.
 
 ## 2.0.0-alpha.3